Rename demo files with .example file extension
